+++ astropy/tests/runner.py
@@ -2,11 +2,13 @@
 
 import inspect
 import os
+import glob
 import copy
 import shlex
 import sys
 import tempfile
 import warnings
+import importlib
 from collections import OrderedDict
 from importlib.util import find_spec
 
@@ -180,7 +182,24 @@ class TestRunnerBase:
         """
 
     def run_tests(self, **kwargs):
-        if not _has_test_dependencies(): # pragma: no cover
+
+        # The following option will include eggs inside a .eggs folder in
+        # sys.path when running the tests. This is possible so that when
+        # runnning python setup.py test, test dependencies installed via e.g.
+        # tests_requires are available here. This is not an advertised option
+        # since it is only for internal use
+        if kwargs.pop('add_local_eggs_to_path', False):
+
+            # Add each egg to sys.path individually
+            for egg in glob.glob(os.path.join('.eggs', '*.egg')):
+                sys.path.insert(0, egg)
+
+            # We now need to force reload pkg_resources in case any pytest
+            # plugins were added above, so that their entry points are picked up
+            import pkg_resources
+            importlib.reload(pkg_resources)
+
+        if not _has_test_dependencies():  # pragma: no cover
             msg = "Test dependencies are missing. You should install the 'pytest-astropy' package."
             raise RuntimeError(msg)